Got accuracy going up when loss goes down!
Switched to L1Loss, which is pretty much just minimizing the sum of our errors (1 for each failure). The tricky part was getting the LR right. There's a lot of oscillation up front, and then it kicks into a smooth, rapid climb when the LR (decreasing linearly) gets down to .001. But a static LR of .001 doesn't work, because it never gets out of its initial unlucky local minimum.